+++ TablePro/Views/Main/Extensions/MainContentCoordinator+Discard.swift
@@ -19,17 +19,7 @@ extension MainContentCoordinator {
         let dbType = connection.type
         var allStatements: [String] = []
 
-        // Add database-specific BEGIN / START TRANSACTION
-        let beginStatement: String
-        switch dbType {
-        case .mysql, .mariadb:
-            beginStatement = "START TRANSACTION"
-        case .mssql:
-            beginStatement = "BEGIN TRANSACTION"
-        default:
-            beginStatement = "BEGIN"
-        }
-        allStatements.append(beginStatement)
+        allStatements.append(dbType.beginTransactionSQL)
 
         // Add user statements
         allStatements.append(contentsOf: statements)